Adjust swap-usage warning level.
Currently it's triggering because gerrit has had a reasonable amount of
activity, and, being java, it's inflated how much memory it's using and
swapped some out. I'm entirely happy with that sequence of events, and
it'll probably happen quite often, so relax the warning level. This
makes it a warning to use more than 100Mb of swap under our current
configuration.

What's really an issue is where some process is swapping serious
amounts, quickly, indicating that some kind of memory unpleasentness
is going on (and thus performance). That's where we want this kind of
alarm which I believe to be not-at-the-current-mark.